edit author, delete author, message, validate

+++ LibraryManagement/Controllers/AuthorsController.cs
@@ -1,45 +1,81 @@
 ﻿using System;
-using System.Collections.Generic;
+using System.Data.Entity;
 using System.Linq;
 using System.Web;
 using System.Web.Mvc;
 using LibraryManagement.Models;
 using PagedList;
 
-namespace LibraryManagement.Controllers
-{
-    public class AuthorsController : Controller
-    {
+namespace LibraryManagement.Controllers {
+    public class AuthorsController : Controller {
         private LibraryEntities _db = new LibraryEntities();
         // GET: Authors
-        public ActionResult Index(int ? page, string searchString)
-        {
-            ViewBag.title = "Library Management - Authors";
-            if (TempData["message"] != null) {
-                ViewBag.message = TempData["message"].ToString();
-            }
+        public ActionResult Index(int? page, string key) {
+            ViewBag.title = "Authors";
             if (page == null) page = 1;
-            var authors = (from a in _db.Authors
-                         select a).OrderBy(a => a.id);
-            if (!String.IsNullOrEmpty(searchString)) {
-                 authors = authors.Where(a => a.author_name.Contains(searchString)).OrderBy(a => a.id);
-                ViewBag.searchValue = searchString;
-            }
-            int pageSize = 6;
             int pageNumber = (page ?? 1);
-            ViewBag.authors = authors;
-            return View(authors.ToPagedList(pageNumber,pageSize));
+            int pageSize = 6;
+            var authors = (from a in _db.Authors
+                           select a).OrderBy(a => a.id);
+            if (!String.IsNullOrEmpty(key)) {
+                authors = authors.Where(a => (a.id + " " + a.author_name).Contains(key)).OrderBy(a => a.id);
+                ViewBag.searchValue = key;
+            }
+            if (authors.Count() == 0) {
+                TempData["message"] = $"Not found anything in system!";
+                TempData["error"] = true;
+            }             
+            return View(authors.ToPagedList(pageNumber, pageSize));
         }
 
         public ActionResult Add() {
+            ViewBag.title = "Authors - Add";
             return View();
         }
 
-        public ActionResult Delete(int authorId) {
-            var author = (from a in _db.Authors where a.id == authorId select a).FirstOrDefault();
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public ActionResult Add(Author author) {
+            if (ModelState.IsValid) {
+                _db.Authors.Add(author);
+                _db.SaveChanges();
+                TempData["message"] = $"Add author successfully!";
+                return RedirectToAction("Index", new { key = author.id + " " + author.author_name });
+            }
+            return View();
+        }
+
+        public ActionResult Edit(int? id) {
+            var author = (from a in _db.Authors where a.id == id select a).SingleOrDefault();
+            if (id == null || author == null) {
+                TempData["message"] = $"Update fail, Cannot found that Author in system!";
+                TempData["error"] = true;
+                return RedirectToAction("Index");
+            }
+            return View(author);
+        }
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public ActionResult Edit(Author author) {
+            if (ModelState.IsValid) {
+                _db.Entry(author).State = EntityState.Modified;
+                _db.SaveChanges();
+                TempData["message"] = $"Update author successfully!";
+                return RedirectToAction("Index", new { key = author.id + " " + author.author_name });
+            }
+               return View(author);
+        }
+
+        public ActionResult Delete(int? id) {
+            var author = (from a in _db.Authors where a.id == id select a).SingleOrDefault();
+            if (id == null || author == null) {
+                TempData["message"] = $"Delete fail, Cannot found that Author in system!";
+                TempData["error"] = true;
+                return RedirectToAction("Index");
+            }
             _db.Authors.Remove(author);
             _db.SaveChanges();
-            TempData["message"] = $"Delete Author {authorId} - {author.author_name} successfully";
+            TempData["message"] = $"Delete Author {id} - {author.author_name} successfully!";
             return RedirectToAction("Index");
         }
     }
